pyproject.toml
[project]
name = "browser-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "browser-use>=0.1.0",
    "jinja2>=3.1.0",
    "langchain-google-genai>=1.0.0",
    "openai>=2.16.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.23.0",
]
